The system for the quizzes rates was changed: now the rates for 1. hard difficulty is 3, 2. for medium is 2, 3. for easy is 1

+++ src/main/java/com/my/quiztaker/forms/UserPublic.java
@@ -2,5 +2,5 @@ package com.my.quiztaker.forms;
 
 public interface UserPublic {
 	String getUsername();
-	Double getRating();
+	Integer getRating();
 }
